gov(gate7): extract certificates_macos_parsers.hpp (end test-drift) + sealed-root error| convention + post-delete verify continue-scan + version bump + not_found doc + obs-log
